scripts/requirements-automation.txt
# Python Automation Scripts Dependencies
python-dateutil==2.8.2
